+++ SharingCodeGatherer/Startup.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
+using Database;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -116,7 +117,7 @@ namespace SharingCodeGatherer
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IServiceProvider services)
         {
             if (env.IsDevelopment())
             {
@@ -141,6 +142,11 @@ namespace SharingCodeGatherer
             });
             #endregion
 
+            // migrate if this is not an inmemory database
+            if (services.GetRequiredService<SharingCodeContext>().Database.ProviderName != "Microsoft.EntityFrameworkCore.InMemory")
+            {
+                services.GetRequiredService<SharingCodeContext>().Database.Migrate();
+            }
         }
     }
 }